refactor: move action menu option to separate file
